First version

+++ src/UseState.js
@@ -1,8 +1,11 @@
 import React from "react";
 
+//Estableciendo el código clave
 const SECURITY_CODE = 'paradigma';
 
+//Función principal
 function UseState (props){
+    //los estados iniciales y setState: un actualizador
     const [state, setState] = React.useState({
         value: '',
         loading: false,
@@ -10,7 +13,8 @@ function UseState (props){
         confirmation: false,
         deleted: false
     })
-
+    //Funciones para que el codigo sea semideclarativo
+    //Actualiza los estados para la pantalla de confirmación
     const onConfirmed = () => {
         setState({
             ...state,
@@ -19,7 +23,7 @@ function UseState (props){
             confirmation: true
         })
     }
-
+    //Actualiza los estados cuando se escribió mal la clave 
     const onError = () =>{
         setState({
             ...state,
@@ -27,24 +31,28 @@ function UseState (props){
             err: true
         })
     }
+    //Actualiza el valor del input: lo que escribe el usuario
     const onWrite = (newValue) => {
         setState({
             ...state,
             value: newValue
         })
     }
+    //Actualiza los estados mientras valida la clave
     const onCheck = () =>{
         setState({
             ...state,
             loading: true
         })
     }
+    //Actualiza los estados para la pantalla de eliminacion
     const onDelete = () =>{
         setState({
             ...state,
             deleted: true
         })
     }
+    //Actualiza los estados para pantalla de restaurado
     const onReset = () => {
         setState({
             ...state,
@@ -54,15 +62,19 @@ function UseState (props){
         })
     }
 
-
+    //Effect: Para que se renderizen ciertos componentes 
+    //unicamente en el estado de carga
     React.useEffect(() => {
 
         if(state.loading){
+            //un timeout para simular un pedido al back-end
             setTimeout(() => {
-
+                
                 if(state.value === SECURITY_CODE){  
+                //Sí la clave es correcta activa la función onConfirmed
                     onConfirmed()
                 } else {
+                //Sí la clave es incorrecta activa la función onError
                     onError()
                 }
 
@@ -71,61 +83,71 @@ function UseState (props){
     // eslint-disable-next-line  
     },[state.loading])
 
+
+    //Tres renders: 
+    //El inicial
    if(!state.deleted && !state.confirmation) {
     return(
         <div>
-                <h2>Hola clase</h2>
-                <p>Eliminar {props.name}</p>
-                {(state.err && !state.loading) && (
-                <p>
-                    Error: por lok
-                </p>)}
-                {(state.loading && 
-                <p>
-                    Cargando...
-                </p>)}
-                <input 
-                    placeholder="Codigo" 
-                    value={state.value}
-                    onChange={(event) => {
-                        onWrite(event.target.value)
-                    }}    
-                />
-                <button 
+            <h2>Usando {props.name}</h2>
+            <p>Semi declarativo compuesto. Usando
+                el Hook useState, objetos y funciones
+            </p>
+                        
+            {(state.err && !state.loading) && (
+                <p>Error: clave incorrecta</p>
+            )}
+            {(state.loading && 
+                <p>Cargando...</p>
+            )}
+            <input 
+                placeholder="Codigo" 
+                value={state.value}
+                onChange={(event) => {
+                    onWrite(event.target.value)
+                }}    
+            />
+            <button 
                 type="submit"
                 onClick={() => { 
                     onCheck()
                 }}
-                >
-                    Comprobar
-                </button>
+            >
+            Comprobar
+            </button>
         </div>
-    )} else if (state.confirmation && !state.deleted){
+    )} // Render de confirmación. Estas seguro?
+    else if (state.confirmation && !state.deleted){
         return(
             <React.Fragment>
+
                 <p>¿Quieres elminarlo?</p>
                 <button
                     onClick={() => {
                         onDelete()
                     }}
-                > SI</button>
+                >SI
+                </button>
                 <button 
                     onClick={() => {
                         onReset()
                     }}
-                > NO</button>
+                >NO
+                </button>
+            
             </React.Fragment>
         )
-    }else {
+    }// Render de eliminado.
+    else {
         return(
             <React.Fragment>
-                <p>se elmino sin exito </p>
+                <p>Se eliminó con éxito </p>
                 <button
-                onClick={() => {
-                  onReset()  
-                }}
+                    onClick={() => {
+                        onReset()
+                    }}
                 >
-                    Resetear
+                Resetear
                 </button>
             </React.Fragment> 
         )
